move initEntities to a helper class to be able to utilise it elsewhere
